move latency compensation to a separate module
