+++ crates/audio/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2024"
 ffmpeg = { workspace = true }
 cpal = { workspace = true }
 tokio.workspace = true
+tracing = { workspace = true }
 
 [target.'cfg(target_os = "macos")'.dependencies]
 cidre = { workspace = true }
